+++ contracts/test/AddressStringUtilTest.sol
@@ -2,10 +2,10 @@
 
 pragma solidity >=0.5.0;
 
-import "../libraries/AddressStringUtil.sol";
+import '../libraries/AddressStringUtil.sol';
 
 contract AddressStringUtilTest {
-    function toAsciiString(address addr, uint len) external pure returns (string memory) {
+    function toAsciiString(address addr, uint256 len) external pure returns (string memory) {
         return AddressStringUtil.toAsciiString(addr, len);
     }
 }
